backend/requirements.txt
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
pillow
django-cors-headers
djoser
psycopg2-binary
